+++ src/test/java/com/softserve/edu/greencity/ui/tests/RegisterComponentTest.java
@@ -3,8 +3,6 @@ package com.softserve.edu.greencity.ui.tests;
 import com.softserve.edu.greencity.ui.data.User;
 import com.softserve.edu.greencity.ui.data.UserRepository;
 import com.softserve.edu.greencity.ui.pages.cabinet.ManualLoginComponent;
-import com.softserve.edu.greencity.ui.pages.cabinet.LoginComponent;
-import com.softserve.edu.greencity.ui.pages.cabinet.LoginManualComponent;
 import com.softserve.edu.greencity.ui.pages.cabinet.ManualRegisterComponent;
 import com.softserve.edu.greencity.ui.pages.cabinet.RegisterComponent;
 import com.softserve.edu.greencity.ui.pages.common.TopGuestComponent;
@@ -194,8 +192,8 @@ public class RegisterComponentTest extends GreenCityTestRunner {
      * //     * @param userLoginCredentials
      * //
      */
-    @Test(dataProvider = "randomValidUserCredentials", enabled = false)
-    public void randomCredsRegistrationLogin(User userLoginCredentials) {
+    @Test(dataProvider = "randomValidUserCredentials")
+    public void randomCredsRegistrationLogin(User userLoginCredentials) throws InterruptedException {
         loadApplication();
         logger.info("Starting randomCredsRegistrationLogin. Input values = "
                 + userLoginCredentials.toString());
@@ -213,14 +211,14 @@ public class RegisterComponentTest extends GreenCityTestRunner {
 
         logger.info("Enter random credentials and temporary email into the form: ");
         manualRegisterComponent.registrationNewRandomUser(userLoginCredentials);
-        Thread.sleep(5000);
+        //Thread.sleep(5000);
        registerComponent.closeRegisterComponentModal();
        ManualLoginComponent manualLoginComponent = new ManualLoginComponent(driver);
-        Thread.sleep(5000);
+        //Thread.sleep(5000);
         manualLoginComponent.inputEmail(userLoginCredentials.getEmail())
                 .inputPassword(userLoginCredentials.getPassword())
-                .clickSignInButton(); // not always success after one click
-Thread.sleep(5000);
+                .clickLoginButton(); // not always success after one click
+//Thread.sleep(5000);
 //        logger.info("get Title curent page: " + driver.getTitle());
 //        Assert.assertEquals(driver.getTitle(), "Home",
 //                "you didn't log in successfully");
